backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib
pdfminer.six
reportlab
spacy
transformers
torch
scikit-learn
pandas

